test(sync): isolate enqueue correlation
